requirements.txt
notion-client
python-dotenv
openai
langchain
tiktoken
chromadb
langchain-community
tenacity
argparse
discord